requirements.txt
# — Core distributed system dependencies —
rpyc==5.3.1                # Remote procedure calls between nodes; allows servers to call methods on each other
etcd3==0.12.0              # Distributed node registry; keeps track of which servers are alive
plyvel==1.5.0              # LevelDB Python bindings; embedded database for storing accounts (no separate server needed)

# — Authentication and crypto —
PyJWT==2.8.0               # JSON Web Tokens (JWTs); cryptographically signed login credentials
cryptography==41.0.7       # TLS certificates and key generation; secures communication between servers

# — Utilities —
requests==2.31.0           # HTTP client; used for IP geolocation in server discovery
tqdm==4.66.1               # Progress bars; shows migration progress in tools/migrate_accounts.py
python-dotenv==1.0.0       # Load environment variables from .env files; simplifies configuration
